[REFACTOR] restructure files to fix #9
Broke down the element_objects.py file into multiple smaller file by groups and moved it into a models directory
